move components into src folder
