+++ bnb-web/src/Funding.tsx
@@ -1,6 +1,6 @@
-import { Progress } from "../components/ui/progress"
-import { Card, CardContent, CardHeader, CardTitle } from "../components/ui/card"
-import { Button } from "../components/ui/button"
+import { Progress } from "./components/ui/progress"
+import { Card, CardContent, CardHeader, CardTitle } from "./components/ui/card"
+import { Button } from "./components/ui/button"
 import { HeartHandshake } from "lucide-react"
 import { Link } from "react-router-dom"
