+++ bnb-web/src/admin/Dashboard.tsx
@@ -1,12 +1,12 @@
 import { Link } from "react-router-dom"
-import { Button } from "../../components/ui/button"
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "../../components/ui/card"
+import { Button } from "../components/ui/button"
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "../components/ui/card"
 import { UserPlus, Package } from 'lucide-react'
 
 export default function AdminPage() {
   return (
     <div className="min-h-screen flex items-center justify-center p-4 bg-gradient-to-b from-background to-background/50">
-      <Card className="max-w-md w-full">
+      <Card className="max-w-md w-full backdrop-blur-sm bg-background/50">
         <CardHeader className="text-center">
           <CardTitle className="text-2xl">Admin Dashboard</CardTitle>
           <CardDescription>Select an administrative function</CardDescription>
