+++ bnb-web/src/components/ui/member-card.tsx
